requirements.txt
flask
boto3
flask-cors
pyjwt
python-dotenv
requests
cryptography
